requirements.txt
albumentations==1.4.0
Flask==3.0.2
Flask_Cors==4.0.0
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.2.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
